Add support for AssumeRole-based credentials
This adds support, when the AWS_ROLE_ARN environment variable, for
metalink to use AssumeRole to fetch temporary credentials for another
account.

Currently, this feature is only supported for AWS. The region is
hard-coded to us-east-1 since it is a required parameter for AssumeRole,
but does not matter when accessing S3 buckets.

More information about AWS AssumeRole functionality:
https://docs.aws.amazon.com/STS/latest/APIReference/API_AssumeRole.html

[#185022387] Add AssumeRole support to metalink