Change format of git+ssh URIs
Previously, you could pass URIs like:
git+ssh://user@host:some-org/some-repo

This was parsed by `net/url` prior to Golang 1.8, even though it is not
actually a valid URL. Post Golang 1.8, you will receive an error when
trying to parse that string as a URL.

Rather than write our own parser, we have changed the format accepted.
You must now pass something like the following (note that there is a
slash now instead of a colon after host):
git+ssh://user@host/some-org/some-repo

This is a parseable URL. The `git.Factory` class reformats this URL so
that it can be passed as the repository arguments to `git clone` within
`git.Source`. This "new" format is actually what was originally
documented in _package.go as an example for using git+ssh scheme.

[#185022387] Add AssumeRole support to metalink

Signed-off-by: Brian Upton <bupton@vmware.com>